+++ src/components/TableModals/ExportModal/index.ts
@@ -1,0 +1,2 @@
+export * from "./ExportModal";
+export { default } from "./ExportModal";
