--- src/components/TableToolbar/Export/index.ts
@@ -1,2 +1,0 @@
-export * from "./Export";
-export { default } from "./Export";
